+++ examples/t07_load_multiple_xml.rs
@@ -1,9 +1,9 @@
 // Copyright © 2025 Stephan Kunz
-
-//! This test implements the seventh tutorial/example from [BehaviorTree.CPP](https://www.behaviortree.dev)
+//! Implements the seventh tutorial/example from [BehaviorTree.CPP](https://www.behaviortree.dev)
+//! using separate xml definitions.
 //!
-//! [tutorial:](https://www.behaviortree.dev/docs/tutorial-basics/tutorial_07_multiple_xml)
-//! [cpp-source:](https://github.com/BehaviorTree/BehaviorTree.CPP/blob/master/examples/t07_load_multiple_xml.cpp)
+//! [tutorial:](https://www.behaviortree.dev/docs/tutorial-basics/tutorial_07_multiple_xml).
+//! [cpp-source:](https://github.com/BehaviorTree/BehaviorTree.CPP/blob/master/examples/t07_load_multiple_xml.cpp).
 
 mod common;