+++ src/behavior/action/mod.rs
@@ -1,8 +1,8 @@
 // Copyright © 2025 Stephan Kunz
-
 //! Built-in `Action` behaviors of [`behaviortree`](crate).
 
 mod change_state_after;
+mod pop_from_queue;
 mod script;
 mod set_blackboard;
 mod sleep;
@@ -10,6 +10,7 @@ mod unset_blackboard;
 
 // flatten
 pub use change_state_after::ChangeStateAfter;
+pub use pop_from_queue::PopFromQueue;
 pub use script::Script;
 pub use set_blackboard::SetBlackboard;
 pub use sleep::Sleep;