+++ tests/actions/mod.rs
@@ -3,4 +3,4 @@
 
 mod blackboard; // set & unset
 //mod sleep;
-//mod pop_from_queue;
+mod pop_from_queue;